requirements.txt
Django==3.1.3
djangorestframework==3.12.2
mysqlclient==2.0.1
django-mysql==3.9
django-cors-headers==3.5.0
#useful when connecting to rabbitMQ
pika==1.1.0
Pillow
psycopg2-binary
Celery
whitenoise==5.2.0
django-heroku==0.3.1